create helper.go
